Add vision handler
